Tests/LibWeb/Text/expected/XHR/XMLHttpRequest-response-header-nonprintable-characters.txt

JSON.stringify(getAllResponseHeaders()) => "heya: , \u000b\f, 1, , , 2\r\n"
JSON.stringify(getResponseHeader("Heya")) => ", \u000b\f, 1, , , 2"
